--- examples/Hedgehog.Xunit.Examples.CSharp/GenAttributeVsContainer/Container.cs
@@ -1,35 +1,0 @@
-﻿namespace Hedgehog.Xunit.Examples.CSharp.GenAttributeVsContainer;
-
-using Hedgehog;
-using Hedgehog.Linq;
-using Hedgehog.Xunit;
-using Gen = Linq.Gen;
-using Range = Linq.Range;
-
-public record PositiveInt(int Value);
-public record NegativeInt(int Value);
-
-public class Generators
-{
-    public static Gen<PositiveInt> GenPositiveInt =>
-        from i in Gen.Int32(Range.Constant(1, int.MaxValue))
-        select new PositiveInt(i);
-
-    public static Gen<NegativeInt> GenNegativeInt =>
-        from i in Gen.Int32(Range.Constant(int.MinValue, -1))
-        select new NegativeInt(i);
-
-    public static AutoGenConfig _ => GenX.defaults
-        .WithGenerator(GenPositiveInt)
-        .WithGenerator(GenNegativeInt);
-}
-
-public class Container
-{
-    [Property(typeof(Generators))]
-    public bool ResultOfAddingPositiveAndNegativeLessThanPositive(
-        PositiveInt positive,
-        NegativeInt negative) =>
-            positive.Value + negative.Value < positive.Value;
-
-}
